LayoutTests/accessibility/win/list-item-role-expected.txt

This test verifies that, on Windows, list items return the "list item" role.

Bug 32688 - MSAA: Accessibility role of list items is wrong

list item
PASS listItem.role is "AXRole: AXTab"
